requirements.backend.txt
fastapi==0.115.11
uvicorn==0.34.0
python-multipart==0.0.20
numpy==2.1.1
opencv-python-headless==4.11.0.86
ultralytics==8.3.89
onnxruntime==1.21.0
onnx==1.17.0
pydantic==2.10.6
pillow==11.1.0